Add client methods for Get, Update resource
This is an initial proof of concept to show the proposed approach of
adding additional methods that use the protection API rather than the
admin endpoints. This is needed for production where admin credentials
won't be available to clients. GetResourceClient(), GetResourcesClient()
and UpdateResourceClient() have been added, and tested with TestGocloak_CreateListGetUpdateDeleteResourceClient()

[Issue: #229]